requirements.txt
matplotlib==3.5.2
Pillow==9.0.1
requests==2.27.1
torch==1.10.0
torchvision==0.11.1
tweepy==4.10.0
